move version to __init__
