+++ imjoy/__init__.py
@@ -1,2 +1,9 @@
 """Package the ImJoy plugin engine."""
-from .const import __version__
+import json
+import pathlib
+
+# read version information from file
+HERE = pathlib.Path(__file__).parent
+VERSION_INFO = json.loads((HERE / "VERSION").read_text())
+__version__ = VERSION_INFO["version"]
+API_VERSION = VERSION_INFO["api_version"]
